abstquiz _fastapi/requirements.txt
boto3==1.34.85
fastapi==0.103.1
pydantic==2.7.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
pytz==2024.1
agraffe==0.7.0
requests==2.31.0
openai==1.23.1
Pillow==10.3.0
# uvicorn==0.29.0